+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-## [0.3.6] 04-01-2022
+## [0.4.0] 05-01-2022
 
 
-### FIX:
+### NEW:
 
-- Fix return type of `Option` to avoid typing errors
+- You can now choose to pass global options to sub commands or not by .
+ The default is that: 
+  - if a processor is passed, we don't pass the options
+  - otherwise, we follow the value of `propagate_options`